level 2 - encrypt and decrypt password using mongoose-encryption package
We will use mongoose-encryption npm package to decrypt/encrypt all the
passport values in the db. this package is automaticlly encrypt the
password when we save a new document(row) to the db, and decrypt it when
we find any document from the db (only by using the moongoose commands)

1. install mongoose-encryption (npm package) and require it in the
   app.js

2.read the documentation of the package and do:
- change userSchema variable from beeing a simple js object to
  mongoose.Schema object

-create a "secret key" string for the decription/incryption

-add to our Schema the encrypt plugin and will specify the
encrypted fields.
--------------------------------
What is the problem to encrypt the password in this way??
if the server hacked and the hacker have the "secret key string"+accsess to the db, he can decrypt all the passwords