level 4 - hashing passwords
level 1-3 were all encryption methods, they used:
password + encryption key to secure a password + cipher method --> Ciphertext
It is easy to encrypt and to decrypt the password, you only need to know the cipher method and the  "encryption key to secure a password"

hashing is no longer need an "encryption key to secure a password ".
becouse of that it is almost imposible to decrypt the password that stored in the db.
password + hash function --> Hash (the hashed function)

So how will it work?
we will never know the user's plain password, we will save the hashed function in the db, and when yhe user would like to sign-in we will hash the sign in and compare it with the saved hash in the db.
we will use npm package called md5