level 5 - salting and hashing passwords with bcrypt.
volnurability to hash passwords :
1. dictionarry attacks
2. hash-table cracks
the same password will always turn into the same hash(hashed password), thereforce if hacker have an acsees to the db were my hashed password is saved, he can use hash-table of (password - and hash result) to reveal the passwords.
one of the most important things to remrmber about hashing is as the number of characters in our passport increase the computation time that it takes to crack the hash increase exponentially.
-----------
Salting:
Salting - is adding a random set of characters to the password and after that hashing it.
in the db we will need to store only the "salt key" and the "hash"(hashed password).

salt is not preventing dictionary attacks
so HOW TO PREVENT dictionary attack?
with alghoritmns as bcrypt that increase the execution time of the hashing.
Salt Rounds - for every time increase in that number the amount of time that it takes to hash the password is doubled.